--- web/src/component/__test__/button.test.tsx
@@ -1,8 +1,0 @@
-import React from "react"
-import { render } from "@testing-library/react"
-import { ShareButton } from "../button"
-
-test("share button", () => {
-  let button = render(<ShareButton></ShareButton>)
-  expect(button).toMatchSnapshot("shareButton")
-})
