+++ web/src/components/__test__/Typograhy.test.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { render } from "@testing-library/react"
-import { H1 } from "../typography"
+import { H1 } from "../Typography"
 
 test("H1", () => {
   let h1 = render(<H1>Testing</H1>)
